Add a .cabal file for making equinox and paradox
Cabal install does not like --main-is too much, and it certianely does
not like a module Main with a function main that is not of type IO ().
Therefore, I make a new file called Runner that copies the old
functionality, and renamed the module names in
{Equinox,Paradox}/Main.hs to main. You should still be able to compile
paradox and equinox by running `make'.